% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attachments.R
\name{add_attachments}
\alias{add_attachments}
\title{Anlagen hinzufügen}
\usage{
add_attachments(directory = NULL, files = NULL, dataset_uid)
}
\arguments{
\item{directory}{Pfad zu Ordner welcher alle Anlagen enthält (Wenn nicht NULL, werden alle Inhalte des Ordners hochgeladen)}

\item{files}{eines oder mehrere Files die als Anlage hochgeladen werden}

\item{dataset_uid}{dataset_uid}
}
\description{
Mit dieser Funktion können eines oder mehere Files bzw. der Inhalt ganzer Ordner als Anlage zu einem Datensatz zugeordnet werden
}
