% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_description_to_field.R
\name{add_description_to_field}
\alias{add_description_to_field}
\title{add_description_to_field}
\usage{
add_description_to_field(dataset_uid, field_name, new_description)
}
\arguments{
\item{dataset_uid}{kann metadata_catalog entnommen werden}

\item{field_name}{Feldname zu dem die Beschreibung gehören soll}

\item{new_description}{neue Beschreibung (wird auch als Label angezeigt)}
}
\value{
wenn erfolgreich: processor_uid, wenn nicht erfolgreich status code
}
\description{
Funktion um Variablenbeschreibungen anzulegen
}
