% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_field_config.R
\name{add_field_config}
\alias{add_field_config}
\title{add_field_config}
\usage{
add_field_config(body, dataset_id)
}
\arguments{
\item{body}{JSON File erzeugt durch create_fields_body}

\item{dataset_id}{kann metadata_catalog entnommen werden}
}
\description{
Funktiion führt POST request mit erzrugtem JSON Body durch
}
