% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata_from_scheme.R
\name{add_metadata_from_scheme}
\alias{add_metadata_from_scheme}
\title{add_metadata_from_scheme}
\usage{
add_metadata_from_scheme(
  filepath,
  harvesting = TRUE,
  zuschreibungen = NULL,
  copy_id = "da_qwsbz2",
  ...
)
}
\arguments{
\item{filepath}{Pfad zum ausgefüllten Schema}

\item{harvesting}{default ist TRUE; wenn kein Harvesting zu opendata.swiss gewünscht wird auf FALSE setzen}

\item{zuschreibungen}{character vevtor mit Zuschreibungen}

\item{copy_id}{Dataset UID des Template Datensatz, in dem die gleichbleibenden Metadaten gespeichert sind}

\item{...}{siehe get_dataset_info}
}
\value{
dataset_uid
}
\description{
Erstellt neuen Datensatz mit entsprechender Kennung und befüllt diesen mit Metadaten
Funktion kann nur auf genau definiertes Schema angewendet werden.
Alle Metadaten im Schema werden auf das Portal gesladen.
}
