% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_resource_to_data.R
\name{add_resource_to_data}
\alias{add_resource_to_data}
\title{add_resource_to_data}
\usage{
add_resource_to_data(resource, title, dataset_uid)
}
\arguments{
\item{resource}{Name der Ressource (kann aus files_list entnommen werden)}

\item{title}{Gewünschter Titel der Ressource}

\item{dataset_uid}{dataset_uid des Datensatz, dem die Ressource zugeordnet werden soll}
}
\description{
Hochegladene Datei (Resoource) mit Datensatz auf ODS verbinden
}
