% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_timeserie_precision.R
\name{add_timeserie_precision}
\alias{add_timeserie_precision}
\title{add_timeserie_precision}
\usage{
add_timeserie_precision(dataset_uid, field_name, annotation_args)
}
\arguments{
\item{dataset_uid}{kann metdadata_catalog entnommen werden}

\item{field_name}{Name des Feldes}

\item{annotation_args}{Possible arguments are year, month and day for date, hour and minute for datetime}
}
\value{
Status Code
}
\description{
add_timeserie_precision
}
