% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_type.R
\name{add_type}
\alias{add_type}
\title{add_type}
\usage{
add_type(dataset_uid, field_name, new_type)
}
\arguments{
\item{dataset_uid}{Kann metadata_catalog entnommen werden}

\item{field_name}{name des Feldes}

\item{new_type}{Neuer Typ (siehe verwendbare Typen unter https://betahelp.opendatasoft.com/management-api/#type)}
}
\description{
add_type
}
