% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_unit.R
\name{add_unit}
\alias{add_unit}
\title{add_unit}
\usage{
add_unit(dataset_uid, field_name, unit)
}
\arguments{
\item{dataset_uid}{kann metadata_catalog entnommen werden}

\item{field_name}{Feld zu dem Unit hinzugefügt werden soll}

\item{unit}{unterstützte Units unter https://betahelp.opendatasoft.com/management-api/#units}
}
\value{
wenn erfolgreich processor_id, wenn nicht Status Code
}
\description{
Einheit zu Variable hinzufügen
}
