% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_user.R
\name{add_user}
\alias{add_user}
\title{add_user}
\usage{
add_user(email_list)
}
\arguments{
\item{email_list}{Muss eine Liste sein, auch wenn nur ein Nuter hinzugefügt werden soll.}
}
\description{
Funktion um ODS-Einladungsmails an einen oder mehrere Mail Adressen zu versenden. Bereits hinzugefügte Nutzer werden nicht erneut eingeladen.
}
