% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_user_to_data.R
\name{add_user_to_data}
\alias{add_user_to_data}
\title{add_user_to_data}
\usage{
add_user_to_data(
  username,
  dataset_uid,
  permission_list = list("explore_restricted_dataset")
)
}
\arguments{
\item{username}{Name des Users (ID)}

\item{dataset_uid}{Datensatz UID}

\item{permission_list}{Liste von Berechtigungen, die dem Nutzer für den Datensatz zugestanden werden (default ist 'explore_restricted_dataset')}
}
\description{
Funktion um User zum Datensatz hinzuzufügen (z.B. Amtsverantwortliche zur Kontrolle vor der Veröffentlichung)
}
