% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_user_to_group.R
\name{add_user_to_group}
\alias{add_user_to_group}
\title{add_user_to_group}
\usage{
add_user_to_group(group_id, user_list)
}
\arguments{
\item{group_id}{Gruppen ID (kann über get_groups ermittelt werden)}

\item{user_list}{Liste von Useramen (nicht Email Adressen). Parameter MUSS eine Liste sein}
}
\description{
Funktion um User zu Gruppe hinzuzufügen
}
