% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_path.R
\name{check_path}
\alias{check_path}
\title{Check Path}
\usage{
check_path()
}
\value{
path_var
}
\description{
Checks whther path is initialised. If not a default path for local catalog will be initialised
}
