% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_resource_with_csv.R
\name{combine_resource_with_csv}
\alias{combine_resource_with_csv}
\title{combine_resource_with_csv}
\usage{
combine_resource_with_csv(
  filepath,
  dataset_uid,
  resource_uid = NULL,
  delim_local_csv = ","
)
}
\arguments{
\item{filepath}{Pfad zum CSV File, das hochgeladen werden soll}

\item{dataset_uid}{dataset_uid des Datensatz, dem die Ressource zugeordnet werden soll}

\item{resource_uid}{resource_id of the resource that should be updated}

\item{delim_local_csv}{seperator for the local csv file (default is ',')}
}
\description{
Funktion um vorhandene Resource mit neuen Daten zu verbinden (rbind). Neue Da
}
