% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{create_dataset}
\usage{
create_dataset(identifier, title, ...)
}
\arguments{
\item{identifier}{technische Kennung}

\item{title}{Titel}

\item{...}{siehe `get_dataset_info`}
}
\value{
dataset_uid
}
\description{
create_dataset
}
