% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_empty_dataset.R
\name{create_empty_dataset}
\alias{create_empty_dataset}
\title{Title}
\usage{
create_empty_dataset(identifier)
}
\arguments{
\item{identifier}{gewünschte Technische Kennung}
}
\value{
dataset_uid des neuen Datensatzes
}
\description{
Funktion um leeren Datensatz zu erstellen. Als Parameter wird die gewünschte Datensatz ID angegeben, die dann auch als Titel verwendet wird
}
