% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fields_body.R
\name{create_fields_body}
\alias{create_fields_body}
\title{create_fields_body}
\usage{
create_fields_body(
  configuration_item,
  edit_field_id,
  new_name = NULL,
  new_label = NULL,
  new_type = NULL,
  new_annotation = NULL,
  annotation_args = NULL,
  new_description = NULL,
  new_order = NULL
)
}
\arguments{
\item{configuration_item}{"rename", "type", "annotate", "description", "order" oder "delete"}

\item{edit_field_id}{Feld das bearbeitet werden soll}

\item{new_name}{Neuer Name ("rename")}

\item{new_label}{Neues Label ("rename")}

\item{new_type}{Neuer Variablentyp ("type")}

\item{new_annotation}{Neue Annotation ("annotate")}

\item{annotation_args}{Argumente für Annotation ("annotate")}

\item{new_description}{Neue Beschreibung einer Variablen ("description)}

\item{new_order}{Neue Reihenfolge ("order")}
}
\value{
JSON Body für entsprechenden API request
}
\description{
Funktion um passenden JSON Body für API Request zu erstellen
}
