% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_dataset_id.R
\name{create_new_dataset_id}
\alias{create_new_dataset_id}
\title{create_new_dataset_id}
\usage{
create_new_dataset_id(test_id, ...)
}
\arguments{
\item{test_id}{String in Form Departement-Amt (Beispiel: sk-stat)}

\item{...}{see `get_dataset_info()`}
}
\value{
dataset_id/technische Kennung
}
\description{
Funktion erzeugt Kennung mit durchgängiger Nummerierung ohne Duplikate
}
