% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_attachements.R
\name{delete_attachements}
\alias{delete_attachements}
\title{delete_attachements}
\usage{
delete_attachements(dataset_uid)
}
\arguments{
\item{dataset_uid}{dataset_uid,dataset_id oder Datensatz Titel können angegeben werden}
}
\description{
Funktion um Datensatz Anhänge zu löschen
}
