% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_dataset.R
\name{delete_dataset}
\alias{delete_dataset}
\title{delete_dataset}
\usage{
delete_dataset(dataset_uid)
}
\arguments{
\item{dataset_uid}{dataset_uid,dataset_id oder Datensatz Titel können angegeben werden}
}
\description{
Funktion um Datensatz zu löschen
}
