% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_field_specification.R
\name{delete_field_specification}
\alias{delete_field_specification}
\title{delete_field_specification}
\usage{
delete_field_specification(dataset_uid, processor_uid)
}
\arguments{
\item{dataset_uid}{dataset_uid des Datensatz, dem die Ressource zugeordnet werden soll}

\item{processor_uid}{UID des zu löschenden field specifications}
}
\description{
Eine  bestimmte Field Specification löschen
}
