% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_resource.R
\name{delete_resource}
\alias{delete_resource}
\title{delete_resource}
\usage{
delete_resource(dataset_uid, resource_uid = NULL)
}
\arguments{
\item{dataset_uid}{dataset_uid des Datensatz, dem die Ressource zugeordnet werden soll}

\item{resource_uid}{resource_id of the resource that should be updated}
}
\description{
Resource eines Datensatzes löschen
}
