% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_resource.R
\name{download_resource}
\alias{download_resource}
\title{download_resource}
\usage{
download_resource(dataset_uid, resource_uid)
}
\arguments{
\item{dataset_uid}{dataset_uid des Datensatz, dem die Ressource zugeordnet werden soll}

\item{resource_uid}{resource_id of the resource that should be updated}
}
\value{
the resource as data.frame
}
\description{
Funktion um Resource eines Datensatzes herunterladen
}
