% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_dataset.R
\name{duplicate_dataset}
\alias{duplicate_dataset}
\title{duplicate_dataset}
\usage{
duplicate_dataset(copy_id, new_id, title)
}
\arguments{
\item{copy_id}{dataset_uid des zu kopierenden Datensatzes}

\item{new_id}{neue technische Kennung}

\item{title}{neuer Titel}
}
\description{
Funktion um Metadaten eines Datensatzes zu kopieren
}
