% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_metadata_from_scheme.R
\name{edit_metadata_from_scheme}
\alias{edit_metadata_from_scheme}
\title{edit_metadata_from_scheme}
\usage{
edit_metadata_from_scheme(
  filepath,
  harvesting = TRUE,
  zuschreibungen = NULL,
  dataset_uid
)
}
\arguments{
\item{filepath}{Pfad zum ausgefüllten Schema}

\item{harvesting}{default ist TRUE; wenn kein Harvesting zu opendata.swiss gewünscht wird auf FALSE setzen}

\item{zuschreibungen}{character vevtor mit Zuschreibungen}

\item{dataset_uid}{Dataset UID des zu bearbeitenden Datensatzes}
}
\description{
Bearbeitet Datensatz mit entsprechender Kennung und ersetzt Metadaten mit Daten aus Excel Schema
Funktion kann nur auf genau definiertes Schema angewendet werden.
Alle Metadaten im Schema werden auf das Portal gesladen.
}
