% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_attachements.R
\name{get_dataset_attachements}
\alias{get_dataset_attachements}
\title{get_dataset_attachements}
\usage{
get_dataset_attachements(dataset_uid)
}
\arguments{
\item{dataset_uid}{Datensatz ID}
}
\value{
Datensatz mit allen Datenquellen die dem entsprechenden ODS Datensatz zugeordnet sind
}
\description{
Alle Anhänge eines Datensatzes anzeigen
}
