% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_changes.R
\name{get_dataset_changes}
\alias{get_dataset_changes}
\title{get_dataset_changes}
\usage{
get_dataset_changes(dataset_uid)
}
\arguments{
\item{dataset_uid}{Datensatz ID}
}
\value{
Datensatz mit allen Datenquellen die dem entsprechenden ODS Datensatz zugeordnet sind
}
\description{
Datensatz Historie herunterladen
}
