% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_status.R
\name{get_dataset_status}
\alias{get_dataset_status}
\title{get_dataset_status}
\usage{
get_dataset_status(dataset_uid)
}
\arguments{
\item{dataset_uid}{Datensatz ID}
}
\value{
Name des dataset_status
}
\description{
Datensatz Status herunterladen
}
