% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_field_processors.R
\name{get_field_processors}
\alias{get_field_processors}
\title{get_field_processors}
\usage{
get_field_processors(dataset_uid)
}
\arguments{
\item{dataset_uid}{kann metadata_catalog entnommen werden}
}
\value{
dataframe mit processor Schritten
}
\description{
Funktion um alle processor Schritte (Umbenennungen, Veränderung Beschreinung, Veränderung Typ,etc.) eines Datensatzes anzuzeigen
}
