% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R
\name{get_groups}
\alias{get_groups}
\title{get_groups}
\usage{
get_groups()
}
\value{
gibt eine Liste mit einem dataframe group_info, einer Liste group_permissions zurück
}
\description{
Funktion um Informationen über die auf ODS verfügbaren Gruppen sowie deren Berechtigungen  zu bekommen.
}
