% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_change.R
\name{get_latest_change}
\alias{get_latest_change}
\title{get_latest_change}
\usage{
get_latest_change(dataset_uid)
}
\arguments{
\item{dataset_uid}{dataset_uid}
}
\value{
change_uid
}
\description{
Neueste change_uid
}
