% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{get_metadata}
\usage{
get_metadata(dataset_uid)
}
\arguments{
\item{dataset_uid}{kann metadata_catalog entnommen werden}
}
\value{
Datensatz mit Metadaten eines gewünschten Datensatzes
}
\description{
Funktion um Metadaten eines bestimmten Datensatzes zu erhalten
}
