% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users.R
\name{get_users}
\alias{get_users}
\title{get_users}
\usage{
get_users()
}
\value{
gibt eine Liste mit einem dataframe user_info, einer Liste user_permissions und einer weiteren liste user_groups zurück
}
\description{
Funktion um Informationen über die auf ODS verfügbaren Nutzer sowie deren Berechtigungen und Gruppenzugehörigkeiten zu bekommen.
}
