% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_resource.R
\name{glimpse_resource}
\alias{glimpse_resource}
\title{glimpse_resource}
\usage{
glimpse_resource(dataset_uid, resource_uid)
}
\arguments{
\item{dataset_uid}{dataset_uid des Datensatz, dem die Ressource zugeordnet werden soll}

\item{resource_uid}{resource_id of the resource that should be updated}
}
\value{
a data.frame
}
\description{
Funktion um ersten paar Zeilen der Resource eines Datensatzes herunterladen
}
