% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_user.R
\name{initialize_user}
\alias{initialize_user}
\title{initialize_user}
\usage{
initialize_user(username = NULL, password = NULL, domain, key = NULL)
}
\arguments{
\item{username}{Username (muss nur angegeben werden, wenn kein Key angegeben wird)}

\item{password}{Password (muss nur angegeben werden, wenn kein Key angegeben wird)}

\item{domain}{Domain der betreffenden Seite z.B. data.tg.ch}

\item{key}{Sollte bereits ein Key erstellt worden sein, der die nötige Berechtigungen enthält, kann dieser hier angeben werden.}
}
\description{
Funktion um User zu initialisieren und neuen Key zu erstellen.
}
