% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invite_user_to_data.R
\name{invite_user_to_data}
\alias{invite_user_to_data}
\title{invite_user_to_data}
\usage{
invite_user_to_data(
  dataset_uid,
  email,
  permission_list = list("explore_restricted_dataset")
)
}
\arguments{
\item{dataset_uid}{Datensatz UID}

\item{email}{Email des Users}

\item{permission_list}{Liste von Berechtigungen. Default ist "explore_restricted_dataset"}
}
\description{
Funktion um User via Email hinzuzufügen und Berechtigungen für bestimmten Datensatz zuzuteilen. Kann vor der Veröffentlichung von OGD
}
