% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_dataset_uid.R
\name{lookup_dataset_uid}
\alias{lookup_dataset_uid}
\title{lookup_dataset_uid}
\usage{
lookup_dataset_uid(id_or_title)
}
\arguments{
\item{id_or_title}{Titel oder dataset_id}
}
\value{
dataset_uid
}
\description{
Funktion um dataset_uid aus dataset_id oder Titel zu ermitteln
}
