% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_all_fields_sortable.R
\name{make_all_fields_sortable}
\alias{make_all_fields_sortable}
\title{make_all_fields_sortable}
\usage{
make_all_fields_sortable(dataset_uid)
}
\arguments{
\item{dataset_uid}{dataset_uid}
}
\description{
Funktion um alle Felder sortierbar zu machen. Funktion wird der Einfachheit halber auf ALLE Felder angewendet. Allerdings werden nur Textfelder verändert. Andere Datentypen werden NICHT verändert.
}
