% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fields_sortable.R
\name{make_fields_sortable}
\alias{make_fields_sortable}
\title{make_fields_sortable}
\usage{
make_fields_sortable(dataset_uid, fields)
}
\arguments{
\item{dataset_uid}{dataset_uid}

\item{fields}{Vektor mit Feldnamen}
}
\description{
Funktion um angegebene Felder sortierbar zu machen. Funktion betrifft nur Text Felder. Andere Felder bleiben unverändert.
}
