% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odsManagementR-package.R
\docType{package}
\name{odsManagementR-package}
\alias{odsManagementR}
\alias{odsManagementR-package}
\title{odsManagementR: ODS Management API Wrapper}
\description{
Mit diesem Package ist es möglich die ODS Management API direkt aus R heraus zu nutzen. Für die Nutzung ist ein Account mit vollen Rechten nötig. Siehe https://betahelp.opendatasoft.com/management-api/.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ogd.tg.ch/}
  \item Report bugs at \url{{{ BugReports }}}
}

}
\author{
\strong{Maintainer}: Felix Lorenz \email{felix.lorenzt@tg.ch}

}
\keyword{internal}
