% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_dataset.R
\name{publish_dataset}
\alias{publish_dataset}
\title{Datensatz publizieren}
\usage{
publish_dataset(dataset_uid)
}
\arguments{
\item{dataset_uid}{kann metadata_catalog entnommen werden}
}
\description{
Funktion published Datensatz, verändert aber NICHT die Sicherheitseinstellung (Datensatz nur für zugelassene Benutzer sichtbar)
}
