% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_field_config.R
\name{put_field_config}
\alias{put_field_config}
\title{put_field_config}
\usage{
put_field_config(body, dataset_id, processor_uid)
}
\arguments{
\item{body}{JSON File erzeugt durch create_fields_body}

\item{dataset_id}{kann metadata_catalog entnommen werden}

\item{processor_uid}{kann field processors entnommen werden}
}
\description{
Funktiion führt PUT request mit erzrugtem JSON Body durch
}
