% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_field.R
\name{rename_field}
\alias{rename_field}
\title{rename_field}
\usage{
rename_field(dataset_uid, old_name, new_name, new_label)
}
\arguments{
\item{dataset_uid}{kann dem metadata_catalog entnommen werden}

\item{old_name}{der Name des Feldes (identifier)}

\item{new_name}{der neue technische Name (sollte dem alten entsprechen)}

\item{new_label}{der neue angezeigte Name (wird als Variablenname angezeigt)}
}
\value{
wenn erfolgreich: processor_uid, wenn nicht erfolgreich status code
}
\description{
rename_field
}
