% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_changes.R
\name{restore_change}
\alias{restore_change}
\title{restore_change}
\usage{
restore_change(dataset_uid, restore_uid)
}
\arguments{
\item{dataset_uid}{Datensatz ID}

\item{restore_uid}{Change ID}
}
\value{
Datensatz mit allen Datenquellen die dem entsprechenden ODS Datensatz zugeordnet sind
}
\description{
Historie wiederherstellen
}
