% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_accrualperiodicity.R
\name{set_accrualperiodicity}
\alias{set_accrualperiodicity}
\title{Set accrual periodicity
Set a value on how often data will be updated}
\usage{
set_accrualperiodicity(dataset_uid, value = NULL)
}
\arguments{
\item{dataset_uid}{dataset_uid}

\item{value}{character for example "Annual" if data is updated every year}
}
\description{
Set accrual periodicity
Set a value on how often data will be updated
}
