% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_metadata.R
\name{set_metadata}
\alias{set_metadata}
\title{set_metadata}
\usage{
set_metadata(dataset_id, template, meta_name, meta_value)
}
\arguments{
\item{dataset_id}{kann metadata_catalog entnommen werden}

\item{template}{kann metadata_catalog entnommen werden}

\item{meta_name}{kann metadata_catalog entnommen werden}

\item{meta_value}{neuer Wert}
}
\description{
Funktion um einen Eintrag in den Metadaten zu ändern
}
