% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_metadata_from_df.R
\name{set_metadata_from_df}
\alias{set_metadata_from_df}
\title{set_metadata_from_df}
\usage{
set_metadata_from_df(dataset_id, metas, title)
}
\arguments{
\item{dataset_id}{dataset_uid}

\item{metas}{Metadaten als Dataframe}

\item{title}{Titel um Kennung als Titel zu überschreiben}
}
\description{
Funktion um Metadaten aus einem ODS Metadaten data.frame zu setzen
}
