% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_title.R
\name{set_title}
\alias{set_title}
\title{set_title}
\usage{
set_title(title, dataset_id)
}
\arguments{
\item{title}{Neuer Titel}

\item{dataset_id}{kann metadata_catalog entnommen werden}
}
\description{
Titel eines Datensatzes setzen
}
