% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUser.R
\name{setUser}
\alias{setUser}
\title{setUser}
\usage{
setUser(username = NULL, password = NULL, apikey, domain, path = NULL)
}
\arguments{
\item{username}{ODS-Username}

\item{password}{ODS-Passwort}

\item{apikey}{API-Key mit den benötigten Berechtigungen (default=NULL)}

\item{domain}{Domain des Portals (z.B. data.tg.ch)}

\item{path}{Pfad wo der Metadata Katalog auf lokal gespeichert werden soll}
}
\description{
Funktion um User Daten zu setzen um package nutzen zu können
}
