% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_attachements.R
\name{update_attachements}
\alias{update_attachements}
\title{update_attachements}
\usage{
update_attachements(directory = NULL, files = NULL, dataset_uid)
}
\arguments{
\item{directory}{Pfad zu Ordner welcher alle Anlagen enthält (Wenn nicht NULL, werden alle Inhalte des Ordners hochgeladen)}

\item{files}{eines oder mehrere Files die als Anlage hochgeladen werden}

\item{dataset_uid}{dataset_uid,dataset_id oder Datensatz Titel können angegeben werden}
}
\description{
Funktion um Datensatz Anhänge zu löschen und neue hochzuladen
}
