% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_group.R
\name{update_group}
\alias{update_group}
\title{update_group}
\usage{
update_group(group_id, group_name = NULL, permission_list = NULL)
}
\arguments{
\item{group_id}{Gruppen ID}

\item{group_name}{Neuer Name}

\item{permission_list}{Liste von permissions (muss als Liste angegeben werden)}
}
\description{
Funktion um Namen und Permissions von Gruppen zu verändern
}
