% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_resource.R
\name{update_resource}
\alias{update_resource}
\title{update_resource}
\usage{
update_resource(resource, title, dataset_uid, resource_uid)
}
\arguments{
\item{resource}{Name der Ressource (kann aus files_list entnommen werden)}

\item{title}{Gewünschter Titel der Ressource}

\item{dataset_uid}{dataset_uid des Datensatz, dem die Ressource zugeordnet werden soll}

\item{resource_uid}{resource_id of the resource that should be updated}
}
\description{
Resource eines Datensatzes ersetzen mit bereits hochgeladenem File
}
