% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_resource_with_csv.R
\name{update_resource_with_csv}
\alias{update_resource_with_csv}
\title{update_resource_with_csv}
\usage{
update_resource_with_csv(filepath, dataset_uid, resource_uid = NULL)
}
\arguments{
\item{filepath}{Pfad zum CSV File, das hochgeladen werden soll}

\item{dataset_uid}{dataset_uid des Datensatz, dem die Ressource zugeordnet werden soll}

\item{resource_uid}{resource_id of the resource that should be updated}
}
\description{
CSV als Resource hochladen und Resource eines bestimmten Datensatzes mit dem hochgeladenen File ersetzen
}
