% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_timeserie_precision.R
\name{update_timeserie_precision}
\alias{update_timeserie_precision}
\title{update_timeserie_precision}
\usage{
update_timeserie_precision(dataset_uid, field_name, annotation_args, proc_id)
}
\arguments{
\item{dataset_uid}{kann metdadata_catalog entnommen werden}

\item{field_name}{Name des Feldes}

\item{annotation_args}{Possible arguments are year, month and day for date, hour and minute for datetime}

\item{proc_id}{processor_uid}
}
\description{
update_timeserie_precision
}
