% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_all.R
\name{upload_all}
\alias{upload_all}
\title{Komplett Upload}
\usage{
upload_all(
  metadata_schema,
  data,
  attachement_directory = NULL,
  attachement_files = NULL,
  ...
)
}
\arguments{
\item{metadata_schema}{Pfad zum Metadaten Schema}

\item{data}{Pfad zu den CSV Daten}

\item{attachement_directory}{siehe `add_attachments`}

\item{attachement_files}{siehe `add_attachments`}

\item{...}{}
}
\value{
dataset_uid
}
\description{
Wrapper um Daten, Metadaten und Anhänge auf ODS hochzuladen
}
