% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file_to_ods.R
\name{upload_file_to_ods}
\alias{upload_file_to_ods}
\title{upload_file_to_ods}
\usage{
upload_file_to_ods(filepath)
}
\arguments{
\item{filepath}{Pfad zum CSV File, das hochgeladen werden soll}
}
\value{
Liste mit Statuscode des API Calls
}
\description{
Funktion um CSV Files hochzuladen
}
