% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_names_to_schema.R
\name{write_names_to_schema}
\alias{write_names_to_schema}
\title{write_names_to_schema (funktioniert nur mit formatierter Schema Excel, keine API Funktion)}
\usage{
write_names_to_schema(filepath, orginal_names_df, new_names_df)
}
\arguments{
\item{filepath}{Pfad zum amtspezifischen Excel File}

\item{orginal_names_df}{Spaltennamen, die das Amt geliefert hat}

\item{new_names_df}{Neue Namen, konfrom mit OGD Leitlinien}
}
\description{
write_names_to_schema (funktioniert nur mit formatierter Schema Excel, keine API Funktion)
}
