% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ogd.R
\name{write_ogd}
\alias{write_ogd}
\title{write_ogd}
\usage{
write_ogd(data, file)
}
\arguments{
\item{data}{R Dataframe}

\item{file}{Pfad zum Speicherort (muss mit .csv enden)}
}
\description{
Funktion um Datensatz mit korrekten OGD Spezifikationen als csv zu speichern.
}
