% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ogd2.R
\name{write_ogd2}
\alias{write_ogd2}
\title{write_ogd2}
\usage{
write_ogd2(data, file, ...)
}
\arguments{
\item{data}{R Dataframe}

\item{file}{Pfad zum Speicherort (muss mit .csv enden)}

\item{...}{see `write.table()`}
}
\description{
Funktion um Datensatz mit korrekten OGD Spezifikationen als csv zu speichern.
UTF-8 wird nicht explizit definiert, da dies manchmal zu Problemen beim Upload auf ODS führt (bei Umlauten)
}
