% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data_to_dataset.R
\name{add_data_to_dataset}
\alias{add_data_to_dataset}
\title{add_data_to_dataset
Funktion um csv file hochzuladen, mit Metadaten zu verknüpfen und Spaltenbeschreibungen zu ergänzen (wrapper)}
\usage{
add_data_to_dataset(
  dataset_uid,
  schema,
  ogd_file,
  resource_title,
  save_names = FALSE,
  dataset_info = NULL
)
}
\arguments{
\item{dataset_uid}{kann metadta_catalog entnommen werden}

\item{schema}{ausgefülltes Schema excel}

\item{ogd_file}{csv file, welches die Daten enthält}

\item{resource_title}{geünschter Titel der Resource auf ODS}

\item{save_names}{boolean der angibt ob die Variablennamen aus dem Excel gespeichert werden sollen (nur bei Automatischem Ablauf benötigt)}

\item{dataset_info}{dataset_info Element (nur bei Automatischem Ablauf benötigt)}
}
\description{
add_data_to_dataset
Funktion um csv file hochzuladen, mit Metadaten zu verknüpfen und Spaltenbeschreibungen zu ergänzen (wrapper)
}
