% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_dataset.R
\name{duplicate_dataset}
\alias{duplicate_dataset}
\title{duplicate_dataset}
\usage{
duplicate_dataset(copy_id, new_id, title, save_local = TRUE)
}
\arguments{
\item{copy_id}{dataset_uid des zu kopierenden Datensatzes}

\item{new_id}{neue technische Kennung}

\item{title}{neuer Titel}

\item{save_local}{siehe `get_dataset_info`}
}
\value{
new_uid dataset_uid des neu erstellten Datensatz
}
\description{
Funktion um Metadaten eines Datensatzes zu kopieren
}
