% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_variables_metadata.R
\name{edit_variables_metadata}
\alias{edit_variables_metadata}
\title{edit_variables_metadata}
\usage{
edit_variables_metadata(dataset_uid, schema)
}
\arguments{
\item{dataset_uid}{kann metadta_catalog entnommen werden}

\item{schema}{ausgefülltes Schema excel}
}
\description{
Funktion um Namen, Spaltenbeschreibungen und Datentypen zu ändern
}
