% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_info.R
\name{get_dataset_info}
\alias{get_dataset_info}
\title{get_dataset_info}
\usage{
get_dataset_info()
}
\description{
Funktion um alle derzeit aufgeschalteten Metadaten herunterzuladen. Hier kann auch die dataset_uid entnommen werden.
Der Catalog wird in der Variable metadata_catalog gespeichert
}
