% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUser.R
\name{setUser}
\alias{setUser}
\title{setUser}
\usage{
setUser(username, password, apikey)
}
\arguments{
\item{username}{ODS-Username}

\item{password}{ODS-Passwort}

\item{apikey}{ODS-API Key}
}
\description{
Funktion um User Daten zu setzen um package nutzen zu können
}
